feat: Add support for cursor and blink control to LCD API
This change introduces the ability to control the cursor and blink features of the LCD display. The `cursor()` and `blink()` methods have been added to the `Lcd` struct, providing a convenient way to toggle these features on and off.

This enhancement allows for more dynamic and visually appealing LCD interactions.